+++ scripts/prepare_sector_network.py
@@ -1839,16 +1839,64 @@ def add_existing_pemmdb_capacities(
         )
 
 
+def _extract_pemmdb_hydro_duration(
+    pemmdb_capacities: pd.DataFrame, tech: str | list[str], year: int
+) -> pd.DataFrame:
+    """
+    Takes a PEMMDB capacities Dataframe and calculates country-wise storage durations for the given hydro technologies.
+
+    Parameters
+    ----------
+    pemmdb_capacities : pd.DataFrame
+        DataFrame containing all PEMMDB capacities.
+    tech : str | list[str]
+        Hydro technology to calculate storage duration for.
+    year : int
+        Planning year for which the values are given.
+
+    Returns
+    -------
+    pd.DataFrame
+        Calculated storage duration max_hours.
+    """
+    if not isinstance(tech, list):
+        tech = [tech]
+
+    caps = (
+        pemmdb_capacities.loc[pemmdb_capacities.carrier.isin(tech)]
+        .query("index_carrier.str.contains('turbine') and unit == 'MW' and p_nom > 0")
+        .groupby("country")
+        .sum()[["p_nom"]]
+        .div(1e3)  # GW
+        .rename(columns={"p_nom": year})
+    )
+    store = (
+        pemmdb_capacities.loc[pemmdb_capacities.carrier.isin(tech)]
+        .query(
+            "index_carrier.str.contains('reservoir') and unit == 'MWh' and e_nom > 0"
+        )
+        .groupby("country")
+        .sum()[["e_nom"]]
+        .div(1e3)  # GWh
+        .rename(columns={"e_nom": year})
+    )
+
+    max_hours = store / caps
+
+    return max_hours
+
+
 def _add_tyndp_scaling_factor(n, pemmdb_capacities, ppl, year):
     """
-    Add a scaling factor to existing PyPSA hydro capacities to approximately match TYNDP input capacities.
+    Add a scaling factor to existing PyPSA hydro capacities and inflows to approximately match TYNDP input capacities.
+    Replace existing storage durations for PHS and hydro with calculated values from PEMMDB input.
     """
 
     # Compute scaling factors
     #########################
 
     # PyPSA hydro capacities
-    pypsa_hydro = (
+    pypsa_hydro_total = (
         ppl.query('carrier == "ror" or carrier == "PHS" or carrier == "hydro"')
         .groupby(["country"])
         .sum()
@@ -1856,7 +1904,7 @@ def _add_tyndp_scaling_factor(n, pemmdb_capacities, ppl, year):
     )
 
     # TYNDP hydro capacities
-    tyndp_hydro = (
+    tyndp_hydro_total = (
         pemmdb_capacities.query(
             "carrier.str.contains('hydro') and index_carrier.str.contains('turbine') and unit == 'MW' and p_nom > 0"
         )
@@ -1867,31 +1915,71 @@ def _add_tyndp_scaling_factor(n, pemmdb_capacities, ppl, year):
     )
 
     # Calculate country-wise scaling factors
-    sf_hydro = pd.concat([tyndp_hydro, pypsa_hydro], axis=1).assign(
+    sf_hydro_total = pd.concat([tyndp_hydro_total, pypsa_hydro_total], axis=1).assign(
         sf=lambda df: df[year] / df.p_nom,
     )[["sf"]]
+
+    # Calculate country-wise max_hours for PHS and reservoirs/pondages
+    phs_max_hours = _extract_pemmdb_hydro_duration(
+        pemmdb_capacities, ["hydro-phs", "hydro-phs-pure"], year
+    )
+    reservoir_max_hours = _extract_pemmdb_hydro_duration(
+        pemmdb_capacities, ["hydro-reservoir", "hydro-pondage"], year
+    )
 
     # Multiply capacities by scaling factor
     #######################################
 
     # ror
+    #####
     ror_i = n.generators.query("carrier.str.contains('ror')").index
-    caps_scaled = n.generators.loc[ror_i, "p_nom"] * n.generators.loc[ror_i, "bus"].str[
-        :2
-    ].map(sf_hydro.sf)
+
+    # Scale capacities
+    caps_scaled = n.generators.loc[ror_i, "p_nom"] * (
+        n.generators.loc[ror_i, "bus"].str[:2].map(sf_hydro_total.sf)
+    )
     n.generators.loc[ror_i, "p_nom"] = caps_scaled
+
     # PHS
+    #####
     phs_i = n.storage_units.query("carrier.str.contains('PHS')").index
-    caps_scaled = n.storage_units.loc[phs_i, "p_nom"] * n.storage_units.loc[
-        phs_i, "bus"
-    ].str[:2].map(sf_hydro.sf)
+
+    # Scale capacities
+    caps_scaled = n.storage_units.loc[phs_i, "p_nom"] * (
+        n.storage_units.loc[phs_i, "bus"].str[:2].map(sf_hydro_total.sf)
+    )
     n.storage_units.loc[phs_i, "p_nom"] = caps_scaled
+
+    # Replace max_hours where available
+    tyndp_max_hours = (
+        n.storage_units.loc[phs_i, "bus"].str[:2].map(phs_max_hours[year]).dropna()
+    )
+    n.storage_units.loc[tyndp_max_hours.index, "max_hours"] = tyndp_max_hours
+
     # hydro
+    #######
     hydro_i = n.storage_units.query("carrier.str.contains('hydro')").index
-    caps_scaled = n.storage_units.loc[hydro_i, "p_nom"] * n.storage_units.loc[
-        hydro_i, "bus"
-    ].str[:2].map(sf_hydro.sf)
+
+    # Scale capacities
+    caps_scaled = n.storage_units.loc[hydro_i, "p_nom"] * (
+        n.storage_units.loc[hydro_i, "bus"].str[:2].map(sf_hydro_total.sf)
+    )
     n.storage_units.loc[hydro_i, "p_nom"] = caps_scaled
+
+    # Replace max_hours where available
+    tyndp_max_hours = (
+        n.storage_units.loc[hydro_i, "bus"]
+        .str[:2]
+        .map(reservoir_max_hours[year])
+        .dropna()
+    )
+    n.storage_units.loc[tyndp_max_hours.index, "max_hours"] = tyndp_max_hours
+
+    # Scale inflows
+    inflow_scaled = n.storage_units_t.inflow.loc[:, hydro_i] * (
+        n.storage_units.loc[hydro_i, "bus"].str[:2].map(sf_hydro_total.sf)
+    )
+    n.storage_units_t.inflow.loc[:, hydro_i] = inflow_scaled
 
 
 def add_ammonia(